feat: add tabs with configurators for tanks and other stores
